fix: cast AVG() to numeric for ROUND() in market_aggregation DAG
PostgreSQL's ROUND() does not accept double precision with a scale
argument. Casting AVG() results to ::numeric resolves the error:
"function round(double precision, integer) does not exist"

+++ dags/dag_market_aggregation.py
@@ -33,9 +33,9 @@ def build_market_summary(**context):
             f.trade_date,
             d.market_type,
             COUNT(DISTINCT f.stock_code) AS total_stocks,
-            ROUND(AVG(d.per), 2) AS avg_per,
-            ROUND(AVG(d.pbr), 2) AS avg_pbr,
-            ROUND(AVG(d.roe), 2) AS avg_roe,
+            ROUND(AVG(d.per)::numeric, 2) AS avg_per,
+            ROUND(AVG(d.pbr)::numeric, 2) AS avg_pbr,
+            ROUND(AVG(d.roe)::numeric, 2) AS avg_roe,
             SUM(f.volume) AS total_volume,
             SUM(d.market_cap) AS total_market_cap
         FROM mart.fact_daily_price f
